starting of dockerization
This step is to remove the need to publish to npm,
the idea is that each module can be build as a docker file and then
linked into where it is needed to compose new runtimes with only the
needed modules.

we can go further to separate the running of the code in a new
container later and pipe the data between containers or even across servers.